Перенес предыдущее ДЗ в отдельную папку.
