+++ app/src/main/java/ru/gressor/concatenator/hw1/moxy/view/MoxyView.java
@@ -1,4 +1,4 @@
-package ru.gressor.concatenator.moxy.view;
+package ru.gressor.concatenator.hw1.moxy.view;
 
 import moxy.MvpView;
 import moxy.viewstate.strategy.AddToEndSingleStrategy;
