+++ app/src/main/java/ru/gressor/concatenator/hw1/standard/view/MainActivity.java
@@ -1,20 +1,18 @@
-package ru.gressor.concatenator.moxy.view;
+package ru.gressor.concatenator.hw1.standard.view;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import android.widget.EditText;
 import android.widget.TextView;
 
-import moxy.MvpAppCompatActivity;
-import moxy.presenter.InjectPresenter;
 import ru.gressor.concatenator.R;
-import ru.gressor.concatenator.moxy.presenter.MoxyPresenter;
+import ru.gressor.concatenator.hw1.standard.presenter.Presenter;
 
-public class MoxyActivity extends MvpAppCompatActivity implements MoxyView {
+public class MainActivity extends AppCompatActivity implements MainView {
+    Presenter presenter;
     EditText editText;
     TextView textView;
-
-    @InjectPresenter
-    MoxyPresenter presenter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -24,6 +22,7 @@ public class MoxyActivity extends MvpAppCompatActivity implements MoxyView {
         textView = findViewById(R.id.textView);
         editText = findViewById(R.id.editText);
 
+        presenter = new Presenter(this);
         findViewById(R.id.button).setOnClickListener((v)->processEditText());
     }
