+++ app/src/main/java/ru/gressor/concatenator/hw1/standard/view/MainView.java
@@ -1,4 +1,4 @@
-package ru.gressor.concatenator.standard.view;
+package ru.gressor.concatenator.hw1.standard.view;
 
 public interface MainView {
     void setTextViewText(String text);
